Close files in Finalize for WebM segmenters.
Now Finalize() will close all the files used by the segmenters.  This
is important to allow WebMMuxer::FireOnMediaEndEvent to get the
correct size of the media file.

This also changes the WebM muxer to use 64-bit numbers for the init
and index ranges to correctly support >4GB files.

b/28194272

+++ packager/media/formats/webm/multi_segment_segmenter.h
@@ -29,8 +29,8 @@ class MultiSegmentSegmenter : public Segmenter {
 
   /// @name Segmenter implementation overrides.
   /// @{
-  bool GetInitRangeStartAndEnd(uint32_t* start, uint32_t* end) override;
-  bool GetIndexRangeStartAndEnd(uint32_t* start, uint32_t* end) override;
+  bool GetInitRangeStartAndEnd(uint64_t* start, uint64_t* end) override;
+  bool GetIndexRangeStartAndEnd(uint64_t* start, uint64_t* end) override;
   /// @}
 
  protected:
